feat(account): show referral activity card on /account
Extracts the three-stat summary (free picks available / friends
paid / credits used) into a shared ReferralActivityCard, then
renders it on both /referrals (where it already lived) and
/account (below the profile form, with a link back to
/referrals for the share controls).

The shared useReferralStatus React Query cache means both mounts
hit at most one network request per stale window.
