+++ frontend/src/app/referrals/ReferralsPageContent.tsx
@@ -2,7 +2,7 @@
 
 import * as React from 'react';
 import Link from 'next/link';
-import { Copy, Gift, Share2 } from 'lucide-react';
+import { Copy, Share2 } from 'lucide-react';
 import { toast } from 'sonner';
 
 import { PageLayout } from '@/components/layout/PageLayout';
@@ -10,6 +10,7 @@ import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
 import { Skeleton } from '@/components/ui/skeleton';
+import { ReferralActivityCard } from '@/components/referrals/ReferralActivityCard';
 import { useReferralStatus } from '@/hooks/useReferralStatus';
 import { ROUTES } from '@/lib/constants';
 
@@ -147,64 +148,8 @@ export function ReferralsPageContent() {
           </CardContent>
         </Card>
 
-        <Card>
-          <CardHeader>
-            <CardTitle className="flex items-center gap-2">
-              <Gift className="h-5 w-5" />
-              Your referral activity
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            {isLoading ? (
-              <div className="grid grid-cols-3 gap-4">
-                {Array.from({ length: 3 }).map((_, i) => (
-                  <Skeleton key={i} className="h-20 w-full" />
-                ))}
-              </div>
-            ) : (
-              <div className="grid grid-cols-3 gap-4 text-center">
-                <Stat
-                  label="Free picks available"
-                  value={status.availableCredits}
-                  emphasis={status.availableCredits > 0}
-                />
-                <Stat label="Friends paid" value={status.qualifiedTotal} />
-                <Stat label="Credits used" value={status.redeemedTotal} />
-              </div>
-            )}
-            <p className="text-muted-foreground mt-4 text-xs">
-              Credits unlock <em>only after</em> a referred friend&apos;s payment is
-              confirmed by an admin. We don&apos;t show who used your code — only the
-              counts.
-            </p>
-          </CardContent>
-        </Card>
+        <ReferralActivityCard />
       </div>
     </PageLayout>
-  );
-}
-
-function Stat({
-  label,
-  value,
-  emphasis = false,
-}: {
-  label: string;
-  value: number;
-  emphasis?: boolean;
-}) {
-  return (
-    <div className="rounded-md border p-3">
-      <div
-        className={
-          emphasis
-            ? 'text-2xl font-bold text-green-600 dark:text-green-400'
-            : 'text-2xl font-bold'
-        }
-      >
-        {value}
-      </div>
-      <div className="text-muted-foreground text-xs">{label}</div>
-    </div>
   );
 }
